+++ Pocket Casts TV App/UI/AppCoordinator.swift
@@ -195,6 +195,7 @@ class AppCoordinator {
     }
 
     func remotePlayPauseToggle() {
+        AnalyticsPlaybackHelper.shared.timestampOfLastRemoteAction = .now
         PlaybackManager.shared.remotePlayPauseToggle()
     }
 }