+++ Pocket Casts TV App/UI/Discover/DiscoverFeaturedPodcastCell.swift
@@ -72,6 +72,7 @@ struct DiscoverFeaturedPodcastCell: View {
                     HStack(spacing: 24) {
                         Button(L10n.tvDiscoverFeaturedPlayLatestEpisode) {
                             Task {
+                                AnalyticsPlaybackHelper.shared.currentSource = .discover
                                 let successPlay = await TVDataManager.shared.playLatestEpisode(of: podcast)
                                 await MainActor.run {
                                     if successPlay {
@@ -87,9 +88,15 @@ struct DiscoverFeaturedPodcastCell: View {
                             Text(L10n.tvDiscoverFeaturedGoToPodcast)
                         }
                         .simultaneousGesture(TapGesture().onEnded {
-                            if let listId, let podcastUuid = podcast.uuid {
+                            guard let podcastUuid = podcast.uuid else {
+                                return
+                            }
+                            if let listId {
                                 DiscoverAnalytics.podcastTapped(listId: listId, podcastUuid: podcastUuid, source: source)
                             }
+                            Analytics.track(.discoverFeaturedPodcastTapped, properties: ["uuid": podcastUuid])
+                            AnalyticsHelper.openedFeaturedPodcast()
+                            DiscoverAnalytics.currentFeaturedPodcast = podcastUuid
                         })
                         .focused($focusedButton, equals: FocusValues.goPodcast)
                     }