+++ Pocket Casts TV App/UI/Player/EpisodePlayerButton.swift
@@ -21,5 +21,5 @@ struct EpisodePlayerButton: View {
 }
 
 #Preview {
-    EpisodePlayerButton(model: EpisodeRowViewModel(episode: MockData.makeStubEpisodes().first!, podcast: MockData.makeStubPodcasts().first!))
+    EpisodePlayerButton(model: EpisodeRowViewModel(episode: MockData.makeStubEpisodes().first!, podcast: MockData.makeStubPodcasts().first!, source: .unknown))
 }